requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
argcomplete==3.5.0
asgiref==3.8.1
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
ci-sdr==0.0.2
click==8.1.7
coloredlogs==15.0.1
ConfigArgParse==1.7
contourpy==1.2.1
ctc_segmentation==1.7.4
cycler==0.12.1
Cython==3.0.11
datasets==2.21.0
decorator==5.1.1
dill==0.3.8
Distance==0.1.3
Django==5.0.7
djangorestframework==3.15.2
editdistance==0.8.1
einops==0.8.0
espnet==202402
espnet-tts-frontend==0.0.3
fast-bss-eval==0.1.3
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
g2p-en==2.1.0
gdown==5.2.0
h5py==3.11.0
huggingface-hub==0.24.6
humanfriendly==10.0
hydra-core==1.3.2
idna==3.7
importlib-metadata==4.13.0
inflect==7.0.0
jaconv==0.4.0
jamo==0.4.1
Jinja2==3.1.4
jiwer==3.0.4
joblib==1.4.2
kaldiio==2.18.0
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.9.2
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.9.2
more-itertools==10.4.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
nltk==3.9.1
numba==0.60.0
numpy==1.23.5
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.19.0
opt-einsum==3.3.0
optimum==1.11.0
packaging==24.1
pandas==2.2.2
parallel-wavegan==0.6.1
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
protobuf==5.27.3
pyarrow==17.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
pypinyin==0.44.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-wpe==0.0.1
pytz==2024.1
pyworld==0.3.4
PyYAML==6.0.2
rapidfuzz==3.9.6
regex==2024.7.24
requests==2.32.3
resampy==0.4.3
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.12.0
sentencepiece==0.1.97
six==1.16.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.4.0
sqlparse==0.5.1
sympy==1.13.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.13.3
tomlkit==0.13.2
torch==2.4.0
torch-complex==0.4.4
torchaudio==2.4.0
tqdm==4.66.5
transformers==4.28.1
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.2
xmltodict==0.13.0
xxhash==3.5.0
yarl==1.9.4
yq==3.4.3
zipp==3.20.0
